Added features
Add audio components
Add main menu
Fix level buttons
Create decorations cells for map denerator (trees, rock)
Add sprite and HP for castle
Optimize area zone for last upgrade bomb tower (create object pool)
Optimize bullet collisions

+++ Assets/Scripts/Scriptable Objects/TowerItem.cs
@@ -5,15 +5,15 @@ namespace TowerDefence
     [CreateAssetMenu(fileName = "Tower", menuName = "Towers/Create Tower")]
     public class TowerItem : ScriptableObject
     {
-        [SerializeField] private int _damage;
-        [SerializeField] private int _range;
-        [SerializeField] private int _speed;
+        [SerializeField] private float _damage;
+        [SerializeField] private float _range;
+        [SerializeField] private float _speed;
         [SerializeField] private int _buyPrice;
         [SerializeField] private int _sellPrice;
 
-        public int Damage { get => _damage; set => _damage = value; }
-        public int Range { get => _range; set => _range = value; }
-        public int Speed { get => _speed; set => _speed = value; }
+        public float Damage { get => _damage; set => _damage = value; }
+        public float Range { get => _range; set => _range = value; }
+        public float Speed { get => _speed; set => _speed = value; }
         public int BuyPrice { get => _buyPrice; set => _buyPrice = value; }
         public int SellPrice { get => _sellPrice; set => _sellPrice = value; }
     }
